requirements.txt
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
tiktoken
selenium
readability-lxml
streamlit
streamlit-tags
openpyxl
groq
google-generativeai
streamlit_lottie
-e .
fastapi
uvicorn
twilio
python-multipart
sentence-transformers
qdrant-client
langchain
langchain-community
langchain-text-splitters
langgraph
langchain-groq